backend/requirements.txt
alembic==1.13.2
argon2-cffi
bcrypt==4.2.0
bleach
blinker==1.9.0
click==8.3.0
exceptiongroup==1.3.0
Flask==3.0.0
Flask-Cors==4.0.0
flask-talisman
gunicorn
greenlet==3.2.4
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
Markdown
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
Pillow
pip==25.3
pluggy==1.6.0
pytest==8.3.3
python-dotenv==1.0.1
setuptools==59.6.0
SQLAlchemy==2.0.36
tomli==2.3.0
typing_extensions==4.15.0
Werkzeug==3.1.3
